alpha 0.4.0.15 Development
Officially added `AreaSystem`. One new properties were added, which is `Labor`. Extractors, landfills, and surfaces are removed from the system, and will be readded in other ones. Next, to introduce vital statistics on not only districts but also map tiles, `Utils.BVHUtils` class was made to help map the buildings to areas, using the Bounding Volume Hierarchy rule; this will also be applied in `POISystem` too. Finally, the export time on the test map was 174ms, which was longer than alpha 0.4.0.7's 36ms. But considering the calculations added in, I think this is acceptable and in fact it is still faster than beta 0.3's 738ms.