alpha 0.4.0.20 Development
The prototype to export GeoTIFF was finally done, and it should support Transverse Mercator and writing in different endianness. However, the time consumed significantly increased from beta 0.3's 3.1s to current state's 6.0s, and the another issue was that the map was mirrored. Aside from the GeoTIFF export, the incorrect easting setting in `Geodata.Transform` was identified and corrected.