alpha 0.4.0.26 Development
The prototype to export .dbf file was done. Several issues was fixed, including: (1) incorrect content length in .shp and .shx files & (2) incorrect point count in .shp file. To test the new methods, ``Area.WriteBoundaryDBF()` was implemented, along with several jobs. Although the number of records was the same between .shp and .dbf files, the order didn't match and this requires the further investigation.